Error Handling

This section describes about error object which this package throws.

Error objects

CANBus.Errors.CANBusError — Type

abstract type CANBusError <: Exception end

Base type of error exception of CANBus.jl

CANBus.Errors.CANBusValueError — Type

struct CANBusValueError <: CANBusError
    msg::String
    context::String
end

This error is thrown when passed value is invalid on specific context.

CANBus.Errors.CANBusOpenError — Type

struct CANBusOpenError <: CANBusError
    msg::String
    device::String
    status::String
end

This error is thrown when interface initialization function.

It's status code(3rd entry) is native error code of device driver.

CANBus.Errors.CANBusIOError — Type

struct CANBusIOError <: CANBusError
    msg::String
    context::String
    device::String
    status::String
end

This error is thrown when interface operation is failed.

It's status code(4th entry) is native error code of device driver.

CANBus.Errors.CANBusNotImplementedError — Type

struct CANBusNotImplementedError <: CANBusError
    msg::String
end

This error is thrown when api is not implemented for specific device.