Internals

classDiagram namespace core{ class InterfaceCfgs class Frames } namespace misc{ class VendorAPIs:::classFocus class CommonAPIs:::classFocus } class ConcreteDevices:::classFocus class Devices:::classFocus <<interface>> Devices Interfaces *.. Devices Interfaces ..> InterfaceCfgs Interfaces ..> Frames Devices <|.. "1..*"ConcreteDevices ConcreteDevices ..> VendorAPIs ConcreteDevices ..> CommonAPIs ConcreteDevices ..> Frames ConcreteDevices ..> InterfaceCfgs classDef classFocus fill:#baffc6,stroke:#29862c;

Devices

Abstract type and functions

CANBus.Interfaces.Devices.AbstractDevice — Type

Base Type of Devices

CANBus.Interfaces.Devices.dev_open — Method

Abstract function for setup and open device.

CANBus.Interfaces.Devices.dev_send — Method

Abstract function for send frame.

CANBus.Interfaces.Devices.dev_recv — Method

Abstract function for receive frame.

CANBus.Interfaces.Devices.dev_close — Method

Abstract function for close device.

CANBus.Interfaces.Devices.dev_status — Method

Abstract function for get status.

ConcreteDevices

Kvaser

CANBus.Interfaces.Devices.KvaserDevices.KvaserDevice — Type

KvaserDevice(handleholder::HandleHolder, time_offset::Float64)

Struct for store Kvaser device handle and time offset.

slcan

CANBus.Interfaces.Devices.SlcanDevices.SlcanDevice — Type

SlcanDevice(port::String, bitrate::Integer)

Struct to store SLCAN device handle and buffer.

SocketCAN

CANBus.Interfaces.Devices.SocketCANDevices.SocketCANDevice — Type

SocketCANDevice(socketholder::SocketHolder)

Struct for store SocketCAN device handle.

Vector

CANBus.Interfaces.Devices.VectorDevices.VectorDevice — Type

VectorDevice(pportHandle::Ref{Vxlapi.XLportHandle}, channelMask::Vxlapi.XLaccess,
    time_offset::Float64, notification_hnd::Ref{Vxlapi.XLhandle})

Struct to store Vector device port handle, channel mask, time offset, and notification handle.

VendorAPIs

Kvaser

CANBus.Interfaces.Devices.KvaserDevices.Canlib — Module

Low level API for Kvaser CANlib SDK

slcan

CANBus.Interfaces.Devices.SlcanDevices.slcandef — Module

Vendor specific defs for slcan

SocketCAN

CANBus.Interfaces.Devices.SocketCANDevices.SocketCAN — Module

Low level API for SocketCAN

Vector

CANBus.Interfaces.Devices.VectorDevices.Vxlapi — Module

Low level API for Vector XL Driver Library

CommonAPIs

SerialHAL

CANBus.misc.SerialHAL — Module

Abstraction layer for serial port communication.

WinWrap

CANBus.misc.WinWrap — Module

Wrappers for Windows system call.

BitTiming

CANBus.misc.BitTiming — Module

Internal module for BitTiming calculation and setting.