Interfaces

Base type

CANBus.Interfaces.AbstractCANInterface — Type

Abstract type for Interfaces.

CANBus.Interfaces.send — Function

send(interface::T<:AbstractCANInterface, frame::AbstractFrame)

Abstract function for send message.

Common behavior of concrete implements:

When send successed, return nothing.

When send failed, throws error.

Classic CAN interfaces can send ONLY Frame

CAN FD interfaces can send both Frame and FDFrame

CANBus.Interfaces.recv — Function

recv(interface::T<:AbstractCANInterface; timeout_s::Real=0)

Abstract function for receive message.

Common behavior of concrete implements:

Default non-blocking.

For blocking receivement, set timeout_s in seconds.

Set timeout_s < 0 for infinite bloking.

When receive successed, returns Frame or FDFrame.

When receive queue is empty, returns nothing.

When fails to receive in other reasons, throws error.

Classic CAN interfaces return only Frame object.

CAN FD interfaces return either Frame or FDFrame object.

CANBus.Interfaces.shutdown — Function

shutdown(interface::T<:AbstractCANInterface)

Abstract function for shutdown interface. Always returns nothing.

Fiilter struct

CANBus.Interfaces.AcceptanceFilter — Type

AcceptanceFilter(code_id, mask)

Struct for accept filter.

If this struct is set to Interface, the id is accepted when <received_id> & mask == code_id & mask or (<received_id> xor code_id) & mask == 0. Those are equivalent.

Kvaser

CANBus.Interfaces.KvaserInterfaces.KvaserInterface — Type

KvaserInterface(channel::Int, bitrate::Int;
    silent::Bool, stdfilter::AcceptanceFilter, extfilter::AcceptanceFilter)

Setup Kvaser interface

channel: channel number in integer.

bitrate: bitrate as bit/s in integer.

silent(optional): listen only flag in bool.

stdfilter(optional): standard ID filter in AcceptanceFilter struct.

extfilter(optional): extended ID filter in AcceptanceFilter struct.

CANBus.Interfaces.KvaserInterfaces.KvaserFDInterface — Type

KvaserFDInterface(channel::Int, bitrate::Int, datarate::Int;
    non_iso::Bool, silent::Bool, stdfilter::AcceptanceFilter, extfilter::AcceptanceFilter)

Setup Kvaser interface for CAN FD.

channel: channel number in integer.

bitrate: bitrate as bit/s in integer.

datarate: datarate as bit/s in integer.

non_iso(optional): use non-iso version of CAN FD

silent(optional): listen only flag in bool.

stdfilter(optional): standard ID filter in AcceptanceFilter struct.

extfilter(optional): extended ID filter in AcceptanceFilter struct.

slcan

slcan does not support hardware filter.

CANBus.Interfaces.SlcanInterfaces.SlcanInterface — Type

slcan0 = SlcanInterface(port::String, bitrate::Integer)

slcan is a CAN over serial protocol by CANable. This version is tested on CANable 2.0.

Note

slcan with FD firmware (b158aa7) is seemd to be always on FD mode, thus there is no pure CAN mode. Therefore, this interface exceptionally receives FDFrame when someone sends that.

port: port name string e.g. COM3 on Windows, /dev/ttyACM0 on Linux.

bitrate: bit rate in bit/s

silent(optional): listen only flag in bool.

CANBus.Interfaces.SlcanInterfaces.SlcanFDInterface — Type

slcan0 = SlcanFDInterface(port::String, bitrate::Integer, datarate::Integer)

slcan is a CAN over serial protocol by CANable. This version is tested on CANable 2.0. This interface supports send CAN FD frame.

port: port name string.

bitrate: bit rate in bit/s

datarate: data rate in bit/s

silent(optional): listen only flag in bool.

SocketCAN

CANBus.Interfaces.SocketCANInterfaces.SocketCANInterface — Type

SocketCANInterface(channel::String; filters::Vector{AcceptanceFilter})

Setup SocketCAN interface.

channel: channel name string, e.g. "can0"

filters(optional): list of filters. experimental.

CANBus.Interfaces.SocketCANInterfaces.SocketCANFDInterface — Type

SocketCANFDInterface(channel::String; filters::Vector{AcceptanceFilter})

Setup SocketCAN for CAN FD.

channel: channel name string, e.g. "can0"

filters(optional): list of filters. experimental.

Vector

CANBus.Interfaces.VectorInterfaces.VectorInterface — Type

VectorInterface(channel::Int, bitrate::Int, appname::String)

Setup Vector interface.

channel: channel number in integer.

bitrate: bitrate as bit/s in integer.

apppname: Application Name string in Vector Hardware Manager.

silent(optional): listen only flag in bool.

stdfilter(optional): standard ID filter in AcceptanceFilter struct.

extfilter(optional): extended ID filter in AcceptanceFilter struct.

CANBus.Interfaces.VectorInterfaces.VectorFDInterface — Type

VectorFDInterface(channel::Int, bitrate::Int, datarate::Int, appname::String)

Setup Vector interface for CAN FD.

channel: channel number in integer.

bitrate: bitrate as bit/s in integer.

datarate: datarate as bit/s in integer.

apppname: Application Name string in Vector Hardware Manager.

non_iso(optional): use non-iso version of CAN FD

silent(optional): listen only flag in bool.

stdfilter(optional): standard ID filter in AcceptanceFilter struct.

extfilter(optional): extended ID filter in AcceptanceFilter struct.