Interfaces

Base type

CAN.Interfaces.AbstractCANInterface — Type

Abstract type for Interfaces.

CAN.Interfaces.send — Function

send(interface::T<:AbstractCANInterface, frame::CANalyze.CANFrame)

Abstract function for send message.

Common behavior of concrete implements:

When send successed, return nothing.

When send failed, throws error.

CAN.Interfaces.recv — Function

recv(interface::T<:AbstractCANInterface)

Abstract function for receive message.

Common behavior of concrete implements:

non-blocking

When receive successed, returns CANalyze.CANFrame.

When receive queue is empty, returns nothing.

When fails to receive in other reasons, throws error.

CAN.Interfaces.shutdown — Function

shutdown(interface::T<:AbstractCANInterface)

Abstract function for shutdown interface. Always returns nothing.

Fiilter struct

CAN.Interfaces.AcceptanceFilter — Type

AcceptanceFilter(code_id, mask)

Struct for accept filter.

If this struct is set to Interface, the id is accepted when <received_id> & mask == code_id & mask or (<received_id> xor code_id) & mask == 0. Those are equivalent.

Kvaser

CAN.Interfaces.KvaserInterfaces.KvaserInterface — Type

KvaserInterface(channel::Int, bitrate::Int;
    silent::Bool, stdfilter::AcceptanceFilter, extfilter::AcceptanceFilter)

Setup Kvaser interface

channel: channel number in integer.

bitrate: bitrate as bit/s in integer.

silent(optional): listen only flag in bool.

stdfilter(optional): standard ID filter in AcceptanceFilter struct.

extfilter(optional): extended ID filter in AcceptanceFilter struct.

CAN.Interfaces.KvaserInterfaces.KvaserFDInterface — Type

KvaserFDInterface(channel::Int, bitrate::Int, datarate::Int;
    non_iso::Bool, silent::Bool, stdfilter::AcceptanceFilter, extfilter::AcceptanceFilter)

Setup Kvaser interface for CAN FD.

channel: channel number in integer.

bitrate: bitrate as bit/s in integer.

datarate: datarate as bit/s in integer.

non_iso(optional): use non-iso version of CAN FD

silent(optional): listen only flag in bool.

stdfilter(optional): standard ID filter in AcceptanceFilter struct.

extfilter(optional): extended ID filter in AcceptanceFilter struct.

SocketCAN

CAN.Interfaces.SocketCANInterfaces.SocketCANInterface — Type

SocketCANInterface(channel::String; filters::Vector{AcceptanceFilter})

Setup SocketCAN interface.

channel: channel name string, e.g. "can0"

filters(optional): list of filters. experimental.

CAN.Interfaces.SocketCANInterfaces.SocketCANFDInterface — Type

SocketCANFDInterface(channel::String; filters::Vector{AcceptanceFilter})

Setup SocketCAN for CAN FD.

channel: channel name string, e.g. "can0"

filters(optional): list of filters. experimental.

Vector

CAN.Interfaces.VectorInterfaces.VectorInterface — Type

VectorInterface(channel::Int, bitrate::Int, appname::String)

Setup Vector interface.

channel: channel number in integer.

bitrate: bitrate as bit/s in integer.

apppname: Application Name string in Vector Hardware Manager.

silent(optional): listen only flag in bool.

stdfilter(optional): standard ID filter in AcceptanceFilter struct.

extfilter(optional): extended ID filter in AcceptanceFilter struct.

CAN.Interfaces.VectorInterfaces.VectorFDInterface — Type

VectorFDInterface(channel::Int, bitrate::Int, datarate::Int, appname::String)

Setup Vector interface for CAN FD.

channel: channel number in integer.

bitrate: bitrate as bit/s in integer.

datarate: datarate as bit/s in integer.

apppname: Application Name string in Vector Hardware Manager.

non_iso(optional): use non-iso version of CAN FD

silent(optional): listen only flag in bool.

stdfilter(optional): standard ID filter in AcceptanceFilter struct.

extfilter(optional): extended ID filter in AcceptanceFilter struct.